Separates app from shuttle boilerplate in main.
App returns a configured serenity client.
